+++ pkg/boilerplate/fixer/transport/testdata/transport_sample_empty.go
@@ -1,0 +1,11 @@
+package testdata
+
+import (
+	"github.com/go-kit/kit/transport/grpc"
+)
+
+type grpcXAXAXAXAServer struct {
+	handleMethodOne grpc.Handler
+	handleMethodTwo grpc.Handler
+	handleGetOdds   grpc.Handler
+}